Plot Functional Space and Chosen Functional Indices

Compute a graphical representation of functional indices. To plot functional indices, functional indices values must have been retrieve through the use of the alpha.fd.multidim function.

alpha.multidim.plot(
  sp_faxes_coord,
  asb_sp_w,
  ind_vect = c("fdis", "fnnd", "feve", "fric", "fdiv", "fori", "fspe"),
  details_list,
  asb_vect,
  fd_ind_values,
  faxes = NULL,
  faxes_nm = NULL,
  range_faxes = c(NA, NA),
  color_bg = "grey95",
  size_sp = 1,
  size_centroid = 1,
  size_centroid_asb2 = 1,
  size_vert = 1,
  size_sp_nm = 3,
  color_sp = "#0072B2",
  color_sp_asb2 = "#D55E00",
  color_sp_gp = "gray80",
  color_segment = "#0072B2",
  color_segment_asb2 = "#CC79A7",
  color_centroid = "#0072B2",
  color_centroid_asb2 = "#D55E00",
  color_vert = NA,
  color_vert_asb2 = NA,
  color_ch = "black",
  color_sp_nm = "black",
  fill_sp = "white",
  fill_sp_asb2 = "white",
  fill_sp_gp = "gray80",
  fill_vert = NA,
  fill_vert_asb2 = NA,
  fill_ch = "white",
  fill_centroid = "#0072B2",
  fill_centroid_asb2 = "#D55E00",
  alpha_ch = 0.3,
  shape_sp_gp = 3,
  shape_sp = 16,
  shape_sp_asb2 = 15,
  shape_vert = NA,
  shape_vert_asb2 = NA,
  shape_centroid = 10,
  shape_centroid_asb2 = 12,
  segment_size = 1,
  segment_size_asb2 = 0.5,
  linetype_segment = "solid",
  linetype_segment_asb2 = "dashed",
  scale_inf = 1,
  scale_sup = 3,
  plot_sp_nm = NULL,
  plot_ch = TRUE,
  fontface_nm = "plain",
  name_file = NULL,
  check_input = TRUE
)

Arguments

sp_faxes_coord
a matrix of species coordinates in a chosen functional space. Species coordinates have been retrieved thanks to tr.cont.fspace or quality.fspaces.

asb_sp_w
a matrix linking weight of species (columns) and a set of assemblages (rows)

ind_vect
a vector of character string of the name of functional indices to plot. Indices names must be written in lower case letters. Possible indices to compute are: "fdis", "feve", "fric", "fdiv", "fori" and "fspe". Default: all the indices are computed.

details_list
a list retrieved with the alpha.fd.multidim function containing information about features used to compute indices.

asb_vect
a vector containing the name of the assemblages to represent. Two assemblages can be represented in each plot if asked by the user.

fd_ind_values
a dataframe containing the numeric value of each computed indices. This dataframe is an output of the alpha.fd.multidim function.

faxes
a vector with names of axes to plot. You can only plot from 2 to 4 axes for graphical reasons: vector length should be between 2 and 4. Default: faxes = NULL (the four first axes will be plotted).

faxes_nm
a vector with axes labels if the user want different axes labels than faxes ones. Default: faxes_nm = faxes (labels will the the same that faxes ones).

range_faxes
a vector with minimum and maximum for values for axes. Note that to have a fair representation of position of species in all plots, all axes must have the same range. Default: faxes_lim = c(NA, NA) (the range is computed according to the range of values among all axes, all axes having the same range).

color_bg
a R color name or an hexadecimal code used to fill plot background. Default: color_bg = "grey95".

size_sp
a numeric value referring to the size of species belonging to the global pool but not the plotted assemblage. Default: size_sp = 1.

size_centroid
a numeric value referring to the size of the centroid point. Used for FDiv, FSpe, FDis plotting. Default: size_centroid = 1.

size_centroid_asb2
a numeric value referring to the size of the centroid point for the second assemblage to plot if there is one. Used for FDiv, FSpe, FDis plotting. Default: size_centroid = 1.

size_vert
a numeric value referring to the size of symbol for vertices Default: size_vert = 1.

size_sp_nm
a numeric value for size of species label. Default: size_sp_nm = 3.

color_sp
a R color name or an hexadecimal code referring to the color of species from the studied assemblage. This color is also used for FRic convex hull color. Default: color_sp = "#0072B2".

color_sp_asb2
a R color name or an hexadecimal code referring to the colour of species from the second assemblage to plot if there is one. This color is also used for FRic convex hull color. Default is: color_sp_asb2 = "#D55E00".

color_sp_gp
a R color name or an hexadecimal code to give to species that do not belong to the studied assemblages. Default is: color_sp_global_pool = "gray80".

color_segment
a R color name or an hexadecimal code referring to the color of segments linking species of the studied assemblage and/or centroid for FDiv, FEve, FSpe, FOri, FNND, FDis computation. Defaut is the same than color_sp ; Default: color_segment = "#0072B2".

color_segment_asb2
a R color name or an hexadecimal code referring to the color of segments linking species and/or centroid of the second assemblage to plot if there is one for FDiv, FEve, FSpe, FOri, FNND, FDis computation. Defaut is the same than color_sp_asb2 ; Default: color_segment = "#0072B2".

color_centroid
a R color name or an hexadecimal code referring to the color of the centroid point. Used for FDiv, FSpe, FDis plotting. Default: color_centroid = 1. Defaut is the same than color_sp ; Default: color_segment = "#0072B2".

color_centroid_asb2
a R color name or an hexadecimal code referring to the color of the centroid point for the second assemblage to plot if there is one. Used for FDiv, FSpe, FDis plotting. Default: color_centroid = 1. Defaut is the same than color_sp_asb2 ; Default: color_segment = "#0072B2".

color_vert
a R color name or an hexadecimal code referring to the color of vertices if plotted. If color_vert = NA, vertices are not plotted (for shapes only defined by color, ie shape inferior to 20. Otherwise fill must also be set to NA). Default: color_vert = NA.

color_vert_asb2
a R color name or an hexadecimal code referring to the color of vertices if plotted for the second assemblage. If color_vert = NA, vertices are not plotted (for shapes only defined by color, ie shape < 20. Otherwise fill must also be set to NA). Default: color_vert_asb2 = NA.

color_ch
a R color name or an hexadecimal code referring to the border of the convex hull filled by the pool of species. Default: color_ch = "black".

color_sp_nm
a R color name or an hexadecimal code referring to the colour of species label. Default: color_sp_nm = "black".

fill_sp
a R color name or an hexadecimal code referring to the colour to fill species symbol (if shape_sp > 20) and the assemblage convex hull. Default: fill_sp = '#0072B2'.

fill_sp_asb2
a R color name or an hexadecimal code referring to the colour to fill species symbol for the second assemblage (if shape_sp superior to 20) and its associated convex hull. Default: fill_sp_asb2 = 'white'.

fill_sp_gp
a R color name or an hexadecimal code referring to the colour to fill symbol for species from the global pool (if shape_sp superior to 20). Default: fill_sp_gp = "grey80".

fill_vert
a character value referring to the color for filling symbol for vertices (if shape_vert >20). If fill = NA and color = NA, vertices are not plotted (if shape_vert superior to 20. Otherwise color_vert = NULL is enough). Default is NA.

fill_vert_asb2
a character value referring to the color for filling symbol for vertices (if shape_vert superior to 20) for the second assemblage. If fill = NA and color = NA, vertices are not plotted (if shape_vert superior to 20. Otherwise color_vert = NA is enough). Default is NA.

fill_ch
a R color name or an hexadecimal code referring to the filling of the convex hull filled by the pool of species. Default is: fill_ch = "white".

fill_centroid
a R color name or an hexadecimal code referring to the filling of the centroid (if shape_centroid superior to 20). Default: fill_centroid = '#0072B2'.

fill_centroid_asb2
a R color name or an hexadecimal code referring to the filling of the centroid of the second assemblage (if shape_centroid superior to 20). Default: fill_centroid = "#D55E00".

alpha_ch
a numeric value for transparency of the filling of the convex hull (0 = high transparency, 1 = no transparency). Default is 0.3.

shape_sp_gp
a numeric value referring to the shape used to plot species belonging to the global pool but not to the studied assemblage(s). Default: shape_sp_global_pool = 3 (horizontal cross).

shape_sp
a numeric value referring to the shape used to plot species belonging to the studied assemblage. Default: shape_sp = 16 (filled circle).

shape_sp_asb2
a numeric value referring to the shape used to plot species belonging to the second assemblage to plot if there is one. Default: shape_sp_asb2 = 15 (filled square).

shape_vert
a numeric value referring to the symbol used to show vertices position if plot_vertices = TRUE. Default is 23 (filled diamond).

shape_vert_asb2
a numeric value referring to the shape used to plot vertices of the second assemblage if vertices should be plotted in a different way than other species. If shape_vert = NA, no vertices plotted. Default: shape_vert_asb2 = NA.

shape_centroid
a numeric value referring to the shape used to plot centroid for the given assemblage. Default: shape_centroid = 10 (horizontal cross inside an empty circle).

shape_centroid_asb2
a numeric value referring to the shape used to plot centroid for the second assemblage to plot if there is one. Default: shape_centroid_asb2 = 12 (horizontal cross inside an empty square).

segment_size
a numeric value referring to the size of the segment used to link species of a given assemblage and centroid. Default: segment_size = 1.

segment_size_asb2
a numeric value referring to the size of the segment used to link species of the second assemblage and centroid. Default: segment_size = 0.5.

linetype_segment
a character string or the associated numeric value referring the type of line used for segments linking species of the studied assemblage and/or centroid for FDiv, FEve, FSpe, FOri, FNND, FDis computation. Default: linetype_segment = "solid".

linetype_segment_asb2
a character string or the associated numeric value referring the type of line used for segments linking species and/or centroid of the second assemblage to plot if there is one. Used for FDiv, FEve, FSpe, FOri, FNND, FDis computation. Default: linetype_segment = "dashed".

scale_inf
a numeric value referring to the minimal size of a point in a plot according to each species's relative weight. Default: scale_inf = 1.

scale_sup
a numeric value referring to the minimal size of a point in a plot according to each species's relative weight. Default: scale_inf = 3.

plot_sp_nm
a vector containing species names that are to be plotted. Default: plot_nm_sp = NULL (no name plotted).

plot_ch
a logical value indicating whether the convex hull shaping the pool of species should be illustrated. If plot_ch = TRUE, convex hull of all species in the multidimensional space described in sp_faxes_coord is computed and its projection in 2D spaces are drawn as polygons. Default: plot_ch = TRUE.

fontface_nm
a character string for font of species labels (e.g. "italic", "bold"). Default: fontface_nm = 'plain'.

name_file
a character string with name of file to save the figure (without extension). Default is 'NULL' which means plot is displayed. If several plots are to be saved (for several indices), then files are named as follow "name_file1", "name_file2"...

check_input
a logical value indicating whether key features the inputs are checked (e.g. class and/or mode of objects, names of rows and/or columns, missing values). If an error is detected, a detailed message is returned. Default: check.input = TRUE.

Value

For the given assemblage, return a list of one patchwork figure per functional indice containing plots for combinations of up to four axes.

Author

Camille Magneville and Sebastien Villeger

Examples

if (FALSE) {
# Load Species*Traits dataframe:
data("fruits_traits", package = "mFD")

# Load Assemblages*Species dataframe:      
data("baskets_fruits_weights", package = "mFD") 

# Load Traits categories dataframe:
data("fruits_traits_cat", package = "mFD") 
 
# Compute functional distance 
sp_dist_fruits <- mFD::funct.dist(sp_tr         = fruits_traits,
                                  tr_cat        = fruits_traits_cat,
                                  metric        = "gower",
                                  scale_euclid  = "scale_center",
                                  ordinal_var   = "classic",
                                  weight_type   = "equal",
                                  stop_if_NA    = TRUE)
  
# Compute functional spaces quality to retrieve species coordinates matrix:
fspaces_quality_fruits <- mFD::quality.fspaces(sp_dist = sp_dist_fruits, 
 maxdim_pcoa         = 10,
 deviation_weighting = "absolute",
 fdist_scaling       = FALSE,
 fdendro             = "average")
 
# Retrieve species coordinates matrix:
sp_faxes_coord_fruits <- fspaces_quality_fruits$details_fspaces$sp_pc_coord

# Compute alpha diversity indices:
alpha_fd_indices_fruits <- mFD::alpha.fd.multidim(
  sp_faxes_coord   = sp_faxes_coord_fruits[ , c("PC1", "PC2", "PC3", "PC4")],
  asb_sp_w         = baskets_fruits_weights, 
  ind_vect         = c("fdis", "fmpd", "fnnd", "feve", "fric", "fdiv", 
                       "fori", "fspe"),
  scaling          = TRUE, 
  check_input      = TRUE, 
  details_returned = TRUE)
  
# Retrieve alpha diversity indices table:
fd_ind_values_fruits <- alpha_fd_indices_fruits$functional_diversity_indices
fd_ind_values_fruits
# Retrieve a list of details needed for plots:
details_list_fruits <- alpha_fd_indices_fruits$details

# Compute the plots:
mFD::alpha.multidim.plot(
 sp_faxes_coord        = sp_faxes_coord_fruits, 
 asb_sp_w              = baskets_fruits_weights,
 ind_vect              = c("fdis", "fnnd", "feve", "fric", "fdiv", "fori", 
                           "fspe"),
 details_list          = details_list_fruits,
 asb_vect              = c("basket_1"),
 fd_ind_values         = fd_ind_values_fruits,
 faxes                 = NULL, 
 faxes_nm              = NULL,
 range_faxes       = c(NA, NA),
 color_bg              = "grey95",
 size_sp               = 1,
 size_centroid         = 1,
 size_centroid_asb2    = 1,
 size_vert             = 1,
 size_sp_nm            = 3,
 color_sp              = "#0072B2", 
 color_sp_asb2         = "#D55E00",
 color_sp_gp           = "gray80",
 color_segment         = "#0072B2",
 color_segment_asb2    = "#CC79A7",
 color_centroid        = '#0072B2',
 color_centroid_asb2   = "#D55E00",
 color_vert            = NA, 
 color_vert_asb2       = NA,
 color_ch              = "black",
 color_sp_nm           = "black",
 fill_sp               = "white",
 fill_sp_asb2          = "white",
 fill_sp_gp            = "gray80",
 fill_vert             = NA,
 fill_vert_asb2        = NA,
 fill_ch               = "white",
 fill_centroid         = '#0072B2',
 fill_centroid_asb2    = "#D55E00",
 alpha_ch              = 0.3,
 shape_sp_gp           = 3,
 shape_sp              = 16,
 shape_sp_asb2         = 15,
 shape_vert            = NA,
 shape_vert_asb2       = NA,
 shape_centroid        = 10,
 shape_centroid_asb2   = 12,
 segment_size          = 1,
 segment_size_asb2     = 0.5,
 linetype_segment      = "solid",
 linetype_segment_asb2 = "dashed",
 scale_inf             = 1, 
 scale_sup             = 3,
 plot_sp_nm            = NULL,
 plot_ch               = TRUE,
 fontface_nm           = "plain",
 name_file             = NULL,
 check_input           = TRUE)
}