Summarize Assemblage x Species data frame

This function computes a summary helping you to picture assemblages. For this function to work, there must be no NA in your asb_sp_w data frame.

asb.sp.summary(asb_sp_w)

Arguments

asb_sp_w
a matrix showing assemblages (rows) composition in species (columns). Note that species names must be the names of rows.

Value

A list with:

asb_sp_w_occ

a matrix with species occurrences in each assemblage.

sp_tot_w

a vector gathering species biomass/abundance per species.

asb_tot_w

a vector gathering total abundance/biomass per assemblage.

asb_sp_richn

a vector gathering species richness per assemblage.

asb_sp_nm

a list gathering the names of species of each assemblage.

Author

Camille Magneville and Sebastien Villeger

Examples

if (FALSE) {
# Load Assemblages x Species Matrix
data('baskets_fruits_weights', package = 'mFD')

# Summarize Assemblages Data
mFD::asb.sp.summary(asb_sp_w = baskets_fruits_weights)
}