Plot background of multidimensional plots

This function creates a ggplot object with customized axes range (same for both), names and background

background.plot(range_faxes, faxes_nm, color_bg)

Arguments

range_faxes
a vector with minimum and maximum values of axes. Note that to have a fair representation of position of species in all plots, they should have the same range. Default: range_faxes = c(NA, NA) (the range is computed according to the range of values among all axes).

faxes_nm
a vector with axes labels for figure.

color_bg
a R color name or an hexadecimal code used to fill plot background. Default: color_bg = "grey95".

Value

A ggplot object plotting background of multidimensional graphs.

Author

Camille Magneville and Sebastien Villeger

Examples

if (FALSE) {
 background <- background.plot(range_faxes = c(-1, 2), 
                               faxes_nm    = c("PC 1", "PC 2"), 
                               color_bg    = "grey90") 
}