Plot FIde index

Plot FIde index given a pair of functional axes for one or several assemblages. It adds the centroid of species for each assemblage and segments showing centroid coordinates on functional axes on the background plot.

fide.plot(
  ggplot_bg,
  asb_sp_coord2D,
  asb_sp_relatw,
  asb_centroid_coord2D,
  plot_sp = TRUE,
  shape_sp,
  color_sp,
  fill_sp,
  shape_centroid,
  size_centroid,
  color_centroid,
  fill_centroid,
  color_segment,
  width_segment,
  linetype_segment
)

Arguments

ggplot_bg
a ggplot object of the plot background retrieved through the background.plot function.

asb_sp_coord2D
a list of matrix (ncol = 2) with coordinates of species present in each assemblage for a given pair of functional axes.

asb_sp_relatw
a list of matrix gathering species relative weight in each assemblage. It can be retrieved through the asb.sp.summary or alpha.fd.multidim

asb_centroid_coord2D
a list (with names as in asb_sp_coord2D) of vectors with coordinates of the centroid of species for each assemblage for a given pair of axes.

plot_sp
a logical value indicating whether species of each assemblage should be plotted or not. Default: plot_sp = TRUE

shape_sp
a numeric value referring to the shape used to plot species belonging to the studied assemblage. Default: shape_sp = 16 (filled circle). If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_sp
a R color name or an hexadecimal code referring to the color of species from the studied assemblage. Default: color_sp = "#0072B2". If several assemblages it should be a vector with names as in asb_sp_coord2D.

fill_sp
a R color name or an hexadecimal code referring to the colour to fill species symbol (if shape_sp > 20). Default: fill_sp = '#0072B2'. If several assemblages it should be a vector with names as in asb_sp_coord2D.

shape_centroid
a numeric value referring to the shape used to plot centroid of the studied assemblage. Default: shape_centroid = 18 (filled diamond). If several assemblages it should be a vector with names as in asb_sp_coord2D.

size_centroid
a numeric value referring to the size of species centroid but not the plotted assemblage. Default: size_sp = 1. If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_centroid
a R color name or an hexadecimal code referring to the color of the species centroid from the studied assemblage. If several assemblages it should be a vector with names as in asb_sp_coord2D.

fill_centroid
a R color name or an hexadecimal code referring to the colour to fill assemblage centroid symbol (if shape_sp > 20). If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_segment
a R color name or an hexadecimal code referring to the color of the segment linking axes and centroid from the studied assemblage. If several assemblages it should be a vector with names as in asb_sp_coord2D.

width_segment
a numeric value referring to the size of the segment linking centroid and functional axes. Default: width_segment = 1. If several assemblages it should be a vector with names as in asb_sp_coord2D.

linetype_segment
a character string referring to the linetype used to draw the segment linking centroid and functional axes. Default: linetype_segment = "dashed. If several assemblages it should be a vector with names as in asb_sp_coord2D.

Value

a ggplot object with FIde index, species and background

Author

Camille Magneville and Sébastien Villéger

Examples

if (FALSE) {
}