Plot FRic index

This function plots FRic index for a given pair of functional axes and one or several assemblages. It adds convex hull(s), points and vertices of 1:N assemblages on the background plot

fric.plot(
  ggplot_bg,
  asb_sp_coord2D,
  asb_vertices_nD,
  plot_sp = TRUE,
  color_ch,
  fill_ch,
  alpha_ch,
  shape_sp,
  size_sp,
  color_sp,
  fill_sp,
  shape_vert,
  size_vert,
  color_vert,
  fill_vert
)

Arguments

ggplot_bg
a ggplot object of the plot background retrieved through the background.plot function.

asb_sp_coord2D
a list of matrix (ncol = 2) with coordinates of species present in each assemblage for a given pair of axes for a given pair of functional axes.

asb_vertices_nD
a list (with names as in asb_sp_coord2D) of vectors with names of species being vertices in n dimensions.

plot_sp
a logical value indicating whether species of each assemblage should be plotted or not. Default: plot_sp = TRUE.

color_ch
a R color name or an hexadecimal code referring to the border of the convex hull filled by the pool of species. Default: color_ch = "black". If several assemblages it should be a vector with names as in asb_sp_coord2D.

fill_ch
a R color name or an hexadecimal code referring to the filling of the convex hull filled by the pool of species. Default is: fill_ch = "white". If several assemblages it should be a vector with names as in asb_sp_coord2D.

alpha_ch
a numeric value for transparency of the filling of the convex hull (0 = high transparency, 1 = no transparency). Default: alpha_ch = 0.3.

shape_sp
a numeric value referring to the shape used to plot species belonging to the studied assemblage. Default: shape_sp = 16 (filled circle). If several assemblages it should be a vector with names as in asb_sp_coord2D.

size_sp
a numeric value referring to the size of species belonging to the plotted assemblage. Default: size_sp = 1. If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_sp
a R color name or an hexadecimal code referring to the color of species from the studied assemblage. This color is also used for FRic convex hull color. Default: color_sp = "#0072B2". If several assemblages it should be a vector with names as in asb_sp_coord2D.

fill_sp
a R color name or an hexadecimal code referring to the colour to fill species symbol (if shape_sp > 20) and the assemblage convex hull. Default: fill_sp = '#0072B2'. If several assemblages it should be a vector with names as in asb_sp_coord2D.

shape_vert
a numeric value referring to the shape used to plot vertices if vertices should be plotted in a different way than other species. If shape_vert = NA, no vertices plotted. Default: shape_vert = NA. If several assemblages it should be a vector with names as in asb_sp_coord2D.

size_vert
a numeric value referring to the size of symbol for vertices Default: size_vert = 1. If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_vert
a R color name or an hexadecimal code referring to the color of vertices if plotted. If color_vert = NA, vertices are not plotted (for shapes only defined by color, ie shape inferior to 20. Otherwise fill must also be set to NA). Default: color_vert = NA. If several assemblages it should be a vector with names as in asb_sp_coord2D.

fill_vert
a character value referring to the color for filling symbol for vertices (if shape_vert >20). If fill = NA and color = NA, vertices are not plotted (if shape_vert superior to 20. Otherwise color_vert = NULL is enough). Default is NA. If several assemblages it should be a vector with names as in asb_sp_coord2D.

Value

a ggplot object plotting background of multidimensional graphs and FRic convex hulls

Note

first convex hull (2D projection) of all assemblages are added to the background ggplot object (ggplot_bg), then species not being vertices in n Dimensions-space and lastly species being vertices are plotted

Author

Camille Magneville and Sébastien Villéger

Examples

if (FALSE) {
# Load Species*Traits dataframe:
data("fruits_traits", package = "mFD")

# Load Assemblages*Species dataframe:      
data("baskets_fruits_weights", package = "mFD") 

# Load Traits categories dataframe:
data("fruits_traits_cat", package = "mFD") 
 
# Compute functional distance 
sp_dist_fruits <- mFD::funct.dist(sp_tr         = fruits_traits,
                                  tr_cat        = fruits_traits_cat,
                                  metric        = "gower",
                                  scale_euclid  = "scale_center",
                                  ordinal_var   = "classic",
                                  weight_type   = "equal",
                                  stop_if_NA    = TRUE)
  
# Compute functional spaces quality to retrieve species coordinates matrix:
fspaces_quality_fruits <- mFD::quality.fspaces(sp_dist = sp_dist_fruits, 
 maxdim_pcoa         = 10,
 deviation_weighting = "absolute",
 fdist_scaling       = FALSE,
 fdendro             = "average")
 
# Retrieve species coordinates matrix:
  sp_faxes_coord_fruits <- 
    fspaces_quality_fruits$details_fspaces$sp_pc_coord
    
 # Retrieve species coordinates matrix for the assemblage "basket_1":
  sp_filter <- mFD::sp.filter(asb_nm          = c("basket_1"), 
                              sp_faxes_coord = sp_faxes_coord_fruits, 
                              asb_sp_w       = baskets_fruits_weight)
  sp_faxes_coord_fruits_b1 <- sp_filter$`species coordinates`
 
 # Reduce it to the two studid axes: PC1 and PC2:
 sp_faxes_coord_fruits_b1_2D <- sp_faxes_coord_fruits_b1[, c("PC1", "PC2")]

# Set faxes limits:
# set range of axes if c(NA, NA):
 range_sp_coord  <- range(sp_faxes_coord_fruits)
 range_faxes_lim <- range_sp_coord + c(-1, 1)*(range_sp_coord_fruits[2] - 
 range_sp_coord_fruits[1]) * 0.05
 
 # Retrieve the background plot:
 ggplot_bg_fruits <- mFD::background.plot(
                               range_faxes = ranges_faxes_lim, 
                               faxes_nm    = c("PC 1", "PC 2"), 
                               color_bg    = "grey90") 
                               
 # Retrieve vertices names:
 vert_nm_fruits <- vertices(sp_faxes_coord_fruits_b1, 
  order_2D = TRUE, check_input = TRUE)
                               
 # Plot in white the convex hull of all fruits species:
 ggplot_fric <- mFD::fric.plot(
           ggplot_bg       = ggplot_bg_fruits,
           asb_sp_coord2D  = list(basket_1 = sp_faxes_coord_fruits_b1_2D),
           asb_vertices_nD = vert_nm_fruits,
           plot_sp         = TRUE,
           color_ch        = "black", 
           fill_ch         = "white", 
           alpha_ch        = 0.3,
           size_sp = 1,
           shape_sp = 16,
           color_sp = "red",
           fill_sp = "red",
           size_vert = 1,
           color_vert = "red",
           fill_vert = "red",
           shape_vert = 16)
 ggplot_fric

}