mFD: Compute and Illustrate the Multiple Facets of Functional Diversity

Computing functional traits-based distances between pairs of species for species gathered in assemblages allowing to build several functional spaces. The package allows to compute functional diversity indices assessing the distribution of species (and of their dominance) in a given functional space for each assemblage and the overlap between assemblages in a given functional space, see: Chao et al. (2018) doi:10.1002/ecm.1343 , Maire et al. (2015) doi:10.1111/geb.12299 , Mouillot et al. (2013) doi:10.1016/j.tree.2012.10.004 , Mouillot et al. (2014) doi:10.1073/pnas.1317625111 , Ricotta and Szeidl (2009) doi:10.1016/j.tpb.2009.10.001 . Graphical outputs are included. Visit the 'mFD' website for more information, documentation and examples.

See also

Author

Maintainer: Camille Magneville camille.magneville@gmail.com (ORCID) [copyright holder]

Authors:

Nicolas Loiseau nicolas.loiseau1@gmail.com (ORCID)

Camille Albouy albouycamille@gmail.com (ORCID)

Nicolas Casajus nicolas.casajus@fondationbiodiversite.fr (ORCID)

Thomas Claverie tclaverie@gmail.com (ORCID)

Arthur Escalas arthur.escalas@gmail.com (ORCID)

Fabien Leprieur fabien.leprieur@umontpellier.fr (ORCID)

Eva Maire emg.maire@gmail.com (ORCID)

David Mouillot david.mouillot@umontpellier.fr (ORCID)

Sebastien Villeger sebastien.villeger@cnrs.fr (ORCID)