Plot species

Plot species for one to n assemblages

sp.plot(
  ggplot_bg,
  asb_sp_coord2D,
  asb_vertices_nD = NULL,
  asb_sp_relatw = NULL,
  size_sp,
  shape_sp,
  color_sp,
  fill_sp,
  size_vert,
  shape_vert,
  color_vert,
  fill_vert,
  limits_relatw = c(0, 1),
  range_size_relatw = c(1, 10)
)

Arguments

ggplot_bg
a ggplot object of the plot background retrieved through the background.plot function.

asb_sp_coord2D
a list of matrix (ncol = 2) with coordinates of species present in each assemblage for a given pair of axes.

asb_vertices_nD
a list (with names as in asb_sp_coord2D) of vectors with names of species being vertices in n dimensions.

asb_sp_relatw
a matrix gathering species relative weight in each assemblage. It can be retrieved through the asb.sp.summary

size_sp
a numeric value referring to the size of species belonging to the plotted assemblage. Default: size_sp = 1. If several assemblages it should be a vector with names as in asb_sp_coord2D.

shape_sp
a numeric value referring to the shape of species belonging to the plotted assemblage. Default: shape_sp = 16. If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_sp
a R color name or an hexadecimal code referring to the color of species from the studied assemblage. This color is also used for FRic convex hull color. Default: color_sp = "#0072B2". If several assemblages it should be a vector with names as in asb_sp_coord2D.

fill_sp
a R color name or an hexadecimal code referring to the colour to fill species symbol (if shape_sp > 20) and the assemblage convex hull. Default: fill_sp = '#0072B2'. If several assemblages it should be a vector with names as in asb_sp_coord2D.

size_vert
a numeric value referring to the size of symbol for vertices Default: size_vert = 1. If several assemblages it should be a vector with names as in asb_sp_coord2D.

shape_vert
a numeric value referring to the shape of symbol for vertices Default: shape_vert = NA. If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_vert
a R color name or an hexadecimal code referring to the color of vertices if plotted. If color_vert = NA, vertices are not plotted (for shapes only defined by color, ie shape inferior to 20. Otherwise fill must also be set to NA). Default: color_vert = NA. If several assemblages it should be a vector with names as in asb_sp_coord2D.

fill_vert
a character value referring to the color for filling symbol for vertices (if shape_vert >20). If fill = NA and color = NA, vertices are not plotted (if shape_vert superior to 20. Otherwise color_vert = NULL is enough). Default is NA. If several assemblages it should be a vector with names as in asb_sp_coord2D.

limits_relatw
a vector of two numbers giving the limits to set for the scale of species relative weights.

range_size_relatw
a vector of two numbers specifying the minimum and the maximum size of the plotting symbol for relative weights.

Value

a ggplot object with species plotted on the background plot

Note

if asb_vertices_nD = NULL, all arguments for vertices are ignored and if asb_sp_relatw != NULL, size_sp and size_vert are ignored.

Examples

if (FALSE) {
}