Merge distance object(s) into a single data frame

This function merges distance object(s) into a single data frame which rows are pairs of elements and column(s) distance metric(s). It stands on the dist_long function.

dist.to.df(list_dist)

Arguments

list_dist
a list of dist object(s). All dist objects should have a name (e.g. name of distance metric) and the same labels (i.e. names of sets between which distance was computed).

Value

A data frame which first and second columns (names x1 and x2) contain names of the 2 sets involved in each pair, and with one column for each dist object (named after its name in list_dist.

Author

Sebastien Villeger

Examples

# Create dist objects: 
dist_A <- round(dist(matrix(runif(1:100), 5, 2, 
                      dimnames = list(letters[1:5], NULL))), 2)
dist_B <- round(dist(matrix(runif(1:100), 5, 2, 
                      dimnames = list(letters[1:5], NULL))), 2)
dist_C <- round(dist(matrix(runif(1:100), 5, 2, 
                      dimnames = list(letters[1:5], NULL))), 2)

# First example with only 1 distance:
dist.to.df(list(dA = dist_A))
#>    x1 x2   dA
#> 1   a  b 0.41
#> 2   a  c 0.33
#> 3   a  d 0.38
#> 4   a  e 0.70
#> 5   b  c 0.74
#> 6   b  d 0.59
#> 7   b  e 0.81
#> 8   c  d 0.52
#> 9   c  e 0.72
#> 10  d  e 1.08

# Second example with 3 distances:
dist.to.df(list(d1 = dist_A, d2 = dist_B, d3 = dist_C))
#>    x1 x2   d1   d2   d3
#> 1   a  b 0.41 0.24 0.48
#> 2   a  c 0.33 0.52 0.30
#> 3   a  d 0.38 0.20 0.20
#> 4   a  e 0.70 0.71 0.47
#> 5   b  c 0.74 0.38 0.73
#> 6   b  d 0.59 0.38 0.31
#> 7   b  e 0.81 0.56 0.61
#> 8   c  d 0.52 0.49 0.50
#> 9   c  e 0.72 0.19 0.41
#> 10  d  e 1.08 0.66 0.56