Plot Fori

This function plots FOri index for a given pair of functional axes and for one or several assemblages. It adds the distance of species from the studied to their nearest neighbour(s) from the global pool.

fori.plot(
  ggplot_bg,
  asb_sp_coord2D,
  asb_sp_relatw,
  asb_nn_pool,
  pool_coord2D,
  plot_pool = TRUE,
  plot_sp = TRUE,
  shape_pool,
  size_pool,
  color_pool,
  fill_pool,
  shape_sp,
  color_sp,
  fill_sp,
  color_segment,
  width_segment,
  linetype_segment
)

Arguments

ggplot_bg
a ggplot object of the plot background retrieved through the background.plot function.

asb_sp_coord2D
a list of matrix (ncol = 2) with coordinates of species present in each assemblage for a given pair of functional axes.

asb_sp_relatw
a list of vector gathering species relative weight in each assemblage. It can be retrieved through the alpha.fd.multidim. If several assemblages it should be a vector with names as in asb_sp_coord2D.

asb_nn_pool
a list gathering for each species of a studied assemblage its nearest neighbour(s) in the global pool. If several assemblages it should be a vector with names as in asb_sp_coord2D.

pool_coord2D
a list of matrix (ncol = 2) with coordinates of species present in the global pool for a given pair of functional axes.

plot_pool
a logical value indicating whether species of each assemblage should be plotted or not. Default: plot_pool = TRUE.

plot_sp
a logical value indicating whether species of each assemblage should be plotted or not. Default: plot_sp = TRUE

shape_pool
a numeric value referring to the shape used to plot species pool. Default: shape_pool = 16(filled circle).

size_pool
a numeric value referring to the size of species belonging to the global pool. Default: size_pool = 1.

color_pool
a R color name or an hexadecimal code referring to the color of the pool. This color is also used for FRic convex hull color. Default: color_pool = "#0072B2".

fill_pool
a R color name or an hexadecimal code referring to the colour to fill species symbol (if shape_sp > 20) and the assemblage convex hull. Default: fill_pool = '#0072B2'.

shape_sp
a numeric value referring to the shape used to plot species belonging to the studied assemblage. Default: shape_sp = 16 (filled circle). If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_sp
a R color name or an hexadecimal code referring to the color of species from the studied assemblage. Default: color_sp = "#0072B2". If several assemblages it should be a vector with names as in asb_sp_coord2D.

fill_sp
a R color name or an hexadecimal code referring to the colour to fill species symbol (if shape_sp > 20). Default: fill_sp = '#0072B2'. If several assemblages it should be a vector with names as in asb_sp_coord2D.

color_segment
a R color name or an hexadecimal code referring to the color of the segment linking nearest neighbors in the global pool. If several assemblages it should be a vector with names as in asb_sp_coord2D.

width_segment
a numeric value referring to the size of the segment linking nearest neighbors in the global pool. Default: width_segment = 1. If several assemblages it should be a vector with names as in asb_sp_coord2D.

linetype_segment
a character string referring to the linetype used to link nearest neighbors in the global pool. Default: linetype_segment = "dashed. If several assemblages it should be a vector with names as in asb_sp_coord2D.

Value

a ggplot object with FOri index

Author

Camille Magneville and Sébastien Villéger

Examples

if (FALSE) {
}