Compute FUSE (Functionally Unique, Specialized and Endangered)

This index takes into account species functional uniqueness (also called Functional Originality), species specialisation and species IUCN status.

fuse(sp_dist, sp_faxes_coord, nb_NN = 5, GE, standGE = FALSE)

Arguments

sp_dist

a dist object provided by funct.dist, daisy or dist.ktab.

sp_faxes_coord

a data frame with the coordinates of the species on a multidimensional space based on a selected number of axes derived from a Principal Coordinate Analysis (PCoA). The species are in rows and the PCOA axes are in column.

nb_NN

a numerical value giving the number of nearest neighbor to consider. Default: nb_NN = 5.

GE

a numerical vector giving the IUCN status rank (DD = NA, LC = 0, NT = 1, VU = 2, EN = 3, CR = 4) or the IUCN extinction probability associated with each status. See Mooers et al. (2008) for further information. For example, DD = NA, LC = 0, NT = 0.1, VU = 0.4, EN = 0.666, and CR = 0.999).

standGE

a logical value to standardize the GE values.

Value

A data frame with species in rows and the following metrics in columns:

FUSE: functionally unique, specialized and endangered (see Pimiento et al. (2020);

FUn_std: functional uniqueness standardized between 0 and 1 (see Mouillot et al. (2013);

FSp_std: functional specialization standardized between 0 and 1 (see Mouillot et al. (2013);

References

Mouillot et al. (2013) Rare species support vulnerable functions in high-diversity ecosystems. PLoS Biology, 11, e1001569.
Pimiento et al. (2020) Functional diversity of marine megafauna in the Anthropocene. Science Advances, 6, eaay7650.
Violle et al. (2007) Let the concept of trait be functional! Oikos, 116, 882-892.

Author

Fabien Leprieur and Camille Albouy

Examples

# Load species traits data:
sp_tr <- read.csv(system.file('extdata', 'data_traits_MMA_ursus.csv', 
  package = 'mFD'), dec = ',', sep = ';', header = TRUE, row.names = 1,
  na.strings='NA')

# Trait compilation and ordination:
dimorphism      <- ordered(sp_tr$dimorphism)
breeding_site   <- ordered(sp_tr$breeding_site)
social_behavior <- ordered(sp_tr$social_behavior)
weight_max      <- log(sp_tr$adult_weight_max)
social_group    <- log(sp_tr$social_group_mean)
 
# Trait Matrix construction:
sp_tr_end <- data.frame(
  main_diet = sp_tr$main_diet, 
  foraging_water_depth = sp_tr$foraging_water_depth,
  foraging_location = sp_tr$foraging_location, 
  fasting_strategy = sp_tr$fasting_strategy,
  female_sexual_maturity = sp_tr$female_sexual_maturity, 
  weaning = sp_tr$weaning,
  gestation = sp_tr$gestation, inter_litter = sp_tr$inter_litter,
  breeding_site = sp_tr$breeding_site, 
  social_group = sp_tr$social_group_mean,
  social_behavior = sp_tr$social_behavior, 
  weight_max = sp_tr$adult_weight_max,
  dimorphism = sp_tr$dimorphism)
  
rownames(sp_tr_end) <- rownames(sp_tr)
 
# Function weigthing vector:
v <- c(0.25, 0.25, 0.25, 0.25, 0.20, 0.20, 0.20, 0.20, 0.20, 0.5, 0.5, 0.5, 
  0.5)
   
# Gower distance calculation:
sp_tr_end$main_diet <- as.factor(sp_tr_end$main_diet)
sp_tr_end$foraging_water_depth <- as.factor(sp_tr_end$foraging_water_depth)
sp_tr_end$foraging_location <- as.factor(sp_tr_end$foraging_location)
sp_tr_end$breeding_site <- as.factor(sp_tr_end$breeding_site)
sp_tr_end$social_behavior <- as.factor(sp_tr_end$social_behavior)

sp_dist_tr <- cluster::daisy(sp_tr_end, metric = c('gower'), 
  type = list(symm = c(4)), weights = v)
  
# Principal coordinate analyses
Pcoa <- ade4::dudi.pco(ade4::quasieuclid(sp_dist_tr), scann = FALSE, 
                       nf = 40)
#> Warning: Zero distance(s)

sp_faxes_coord <- Pcoa$li[1:40]
 
# FUSE calculation:
 FUSE_res <- mFD::fuse(
    sp_dist        = sp_dist_tr, 
    sp_faxes_coord = as.matrix(sp_faxes_coord), 
    nb_NN          = 5,  
    GE             = sp_tr$IUCN_num,
    standGE        = TRUE)
 FUSE_res
#>                                   FUSE        FUGE        FSGE      FUn_std
#> Arctocephalus_australis     0.00000000 0.000000000 0.000000000 0.4066626897
#> Arctocephalus_forsteri      0.00000000 0.000000000 0.000000000 0.5170579368
#> Arctocephalus_galapagoensis 0.65704590 0.350268727 0.306777175 0.5592652556
#> Arctocephalus_gazella       0.00000000 0.000000000 0.000000000 0.6208798411
#> Arctocephalus_philippii     0.00000000 0.000000000 0.000000000 0.2779525182
#> Arctocephalus_pusillus      0.00000000 0.000000000 0.000000000 0.2442786896
#> Arctocephalus_townsendi     0.00000000 0.000000000 0.000000000 0.5852680899
#> Arctocephalus_tropicalis    0.00000000 0.000000000 0.000000000 0.2297455947
#> Balaena_mysticetus          0.00000000 0.000000000 0.000000000 0.7452467855
#> Balaenoptera_acutorostrata  0.00000000 0.000000000 0.000000000 0.4460157412
#> Balaenoptera_bonaerensis    0.13238904 0.053165961 0.079223082 0.2184186166
#> Balaenoptera_borealis       0.41103021 0.148801451 0.262228760 0.2139234148
#> Balaenoptera_edeni          0.00000000 0.000000000 0.000000000 0.4661991166
#> Balaenoptera_musculus       0.80941909 0.386580600 0.422838494 0.6292520413
#> Balaenoptera_omurai                 NA          NA          NA 0.2386976599
#> Balaenoptera_physalus       0.56665878 0.235274614 0.331384162 0.3536749038
#> Berardius_arnuxii                   NA          NA          NA 0.3691785951
#> Berardius_bairdii                   NA          NA          NA 0.3306752012
#> Callorhinus_ursinus         0.48846635 0.223661233 0.264805112 0.5012945401
#> Caperea_marginata           0.00000000 0.000000000 0.000000000 0.1832726206
#> Cephalorhynchus_commersonii 0.00000000 0.000000000 0.000000000 0.1835307609
#> Cephalorhynchus_eutropia    0.01817766 0.001748160 0.016429499 0.0069987566
#> Cephalorhynchus_heavisidii  0.01830039 0.002077123 0.016223267 0.0083171263
#> Cephalorhynchus_hectori     0.01749032 0.005307888 0.012182432 0.0070959999
#> Cystophora_cristata         0.45529572 0.257164551 0.198131173 0.5865158316
#> Delphinapterus_leucas       0.00000000 0.000000000 0.000000000 0.6739629047
#> Delphinus_capensis                  NA          NA          NA 0.0264311644
#> Delphinus_delphis           0.00000000 0.000000000 0.000000000 0.3984163792
#> Dugong_dugon                0.38031579 0.238192003 0.142123788 0.5379056252
#> Enhydra_lutris              0.71394550 0.422827681 0.291117820 0.7016950229
#> Erignathus_barbatus         0.00000000 0.000000000 0.000000000 0.5531416419
#> Eschrichtius_robustus       0.00000000 0.000000000 0.000000000 0.8271900919
#> Eubalaena_australis         0.00000000 0.000000000 0.000000000 0.3144018773
#> Eubalaena_glacialis         0.53448288 0.191282690 0.343200186 0.2810689133
#> Eubalaena_japonica          0.64071030 0.299831830 0.340878468 0.4661757614
#> Eumetopias_jubatus          0.23895925 0.106981545 0.131977702 0.4516548620
#> Feresa_attenuata            0.00000000 0.000000000 0.000000000 0.2955099743
#> Globicephala_macrorhynchus  0.00000000 0.000000000 0.000000000 0.5255589486
#> Globicephala_melas          0.00000000 0.000000000 0.000000000 0.4990595861
#> Grampus_griseus             0.00000000 0.000000000 0.000000000 0.5233269607
#> Halichoerus_grypus          0.00000000 0.000000000 0.000000000 0.1032026841
#> Histriophoca_fasciata       0.00000000 0.000000000 0.000000000 0.6521823871
#> Hydrurga_leptonyx           0.00000000 0.000000000 0.000000000 0.6622353793
#> Hyperoodon_ampullatus               NA          NA          NA 0.2547878402
#> Hyperoodon_planifrons       0.00000000 0.000000000 0.000000000 0.2557952131
#> Indopacetus_pacificus               NA          NA          NA 0.2850555343
#> Kogia_breviceps                     NA          NA          NA 0.2839209537
#> Kogia_sima                          NA          NA          NA 0.1456685909
#> Lagenodelphis_hosei         0.00000000 0.000000000 0.000000000 0.7241772659
#> Lagenorhynchus_acutus       0.00000000 0.000000000 0.000000000 0.4544319444
#> Lagenorhynchus_albirostris  0.00000000 0.000000000 0.000000000 0.0907709847
#> Lagenorhynchus_australis    0.00000000 0.000000000 0.000000000 0.0010393047
#> Lagenorhynchus_cruciger     0.00000000 0.000000000 0.000000000 0.0020371106
#> Lagenorhynchus_obliquidens  0.00000000 0.000000000 0.000000000 0.3327438642
#> Lagenorhynchus_obscurus     0.00000000 0.000000000 0.000000000 0.4240208007
#> Leptonychotes_weddellii     0.00000000 0.000000000 0.000000000 0.5404210014
#> Lissodelphis_borealis       0.00000000 0.000000000 0.000000000 0.5812804661
#> Lissodelphis_peronii        0.00000000 0.000000000 0.000000000 0.2947422923
#> Lobodon_carcinophaga        0.00000000 0.000000000 0.000000000 0.5833807958
#> Lontra_felina               0.59408663 0.358536525 0.235550101 0.5749777422
#> Megaptera_novaeangliae      0.00000000 0.000000000 0.000000000 0.3564239828
#> Mesoplodon_bidens                   NA          NA          NA 0.2401439269
#> Mesoplodon_bowdoini                 NA          NA          NA 0.0644576583
#> Mesoplodon_carlhubbsi               NA          NA          NA 0.0008237359
#> Mesoplodon_densirostris             NA          NA          NA 0.2058523276
#> Mesoplodon_europaeus                NA          NA          NA 0.1358313979
#> Mesoplodon_ginkgodens               NA          NA          NA 0.0013549553
#> Mesoplodon_grayi                    NA          NA          NA 0.0544755693
#> Mesoplodon_hectori                  NA          NA          NA 0.0533639848
#> Mesoplodon_layardii                 NA          NA          NA 0.0156247025
#> Mesoplodon_mirus                    NA          NA          NA 0.2615344649
#> Mesoplodon_perrini                  NA          NA          NA 0.0007873510
#> Mesoplodon_peruvianus               NA          NA          NA 0.0022081471
#> Mesoplodon_stejnegeri               NA          NA          NA 0.0019316220
#> Mesoplodon_traversii                NA          NA          NA 0.0000000000
#> Mirounga_angustirostris     0.00000000 0.000000000 0.000000000 1.0000000000
#> Mirounga_leonina            0.00000000 0.000000000 0.000000000 0.6948598959
#> Monachus_monachus           0.25265231 0.138305521 0.114346785 0.1977684465
#> Monachus_schauinslandi      0.69523452 0.360695808 0.334538714 0.5791027790
#> Monodon_monoceros           0.00000000 0.000000000 0.000000000 0.7032597866
#> Neophoca_cinerea            0.36345229 0.102217314 0.261234976 0.1434988631
#> Neophocaena_phocaenoides    0.32176325 0.239830220 0.081933034 0.5420666711
#> Odobenus_rosmarus                   NA          NA          NA 0.5803602059
#> Ommatophoca_rossii          0.00000000 0.000000000 0.000000000 0.6659562055
#> Orcaella_brevirostris       0.04133240 0.007567472 0.033764930 0.0101282373
#> Orcaella_heinsohni          0.05624410 0.014527933 0.041716166 0.0292679517
#> Orcinus_orca                        NA          NA          NA 0.7536273607
#> Otaria_flavescens           0.00000000 0.000000000 0.000000000 0.0808815948
#> Pagophilus_groenlandicus    0.00000000 0.000000000 0.000000000 0.7570041744
#> Peponocephala_electra       0.00000000 0.000000000 0.000000000 0.5314927079
#> Phoca_largha                0.00000000 0.000000000 0.000000000 0.5982829480
#> Phoca_vitulina              0.00000000 0.000000000 0.000000000 0.4543455128
#> Phocarctos_hookeri          0.29111541 0.057001064 0.234114342 0.0782092491
#> Phocoena_dioptrica          0.00000000 0.000000000 0.000000000 0.2513166389
#> Phocoena_phocoena           0.00000000 0.000000000 0.000000000 0.0736635094
#> Phocoena_sinus              0.30989600 0.170988841 0.138907159 0.1864775088
#> Phocoena_spinipinnis        0.11003937 0.064171975 0.045867398 0.2651030237
#> Phocoenoides_dalli          0.00000000 0.000000000 0.000000000 0.2916871646
#> Physeter_macrocephalus      0.54891552 0.320788106 0.228127414 0.7564270292
#> Pontoporia_blainvillei      0.27464151 0.170199056 0.104442451 0.3710816347
#> Pseudorca_crassidens        0.17203888 0.112089828 0.059949053 0.4744533574
#> Pusa_caspica                0.36513400 0.135912706 0.229221292 0.1941091818
#> Pusa_hispida                0.00000000 0.000000000 0.000000000 0.1823213934
#> Sotalia_guianensis          0.01946782 0.013573402 0.005894418 0.0546637568
#> Sousa_chinensis             0.20619793 0.131538281 0.074659645 0.5623262797
#> Sousa_teuszii               0.38826398 0.241129462 0.147134521 0.2726857890
#> Stenella_attenuata          0.00000000 0.000000000 0.000000000 0.4500796407
#> Stenella_clymene            0.00000000 0.000000000 0.000000000 0.3747462296
#> Stenella_coeruleoalba       0.00000000 0.000000000 0.000000000 0.2311412497
#> Stenella_frontalis          0.00000000 0.000000000 0.000000000 0.3490478318
#> Stenella_longirostris       0.00000000 0.000000000 0.000000000 0.5133002403
#> Steno_bredanensis           0.00000000 0.000000000 0.000000000 0.1334810508
#> Tasmacetus_shepherdi                NA          NA          NA 0.3474343349
#> Trichechus_manatus          0.33814860 0.224659317 0.113489282 0.5037922863
#> Trichechus_senegalensis     0.36824108 0.235678375 0.132562708 0.5315342856
#> Tursiops_aduncus            0.04686502 0.012864309 0.034000707 0.0517896426
#> Tursiops_truncatus          0.00000000 0.000000000 0.000000000 0.3052280508
#> Ursus_maritimus             0.67482557 0.374675485 0.300150081 0.9090386502
#> Zalophus_californianus      0.00000000 0.000000000 0.000000000 0.1204667495
#> Zalophus_wollebaeki         0.32738815 0.054693281 0.272694870 0.0749554707
#> Ziphius_cavirostris         0.00000000 0.000000000 0.000000000 0.1383689569
#>                                FSp_std
#> Arctocephalus_australis     0.44574291
#> Arctocephalus_forsteri      0.51871477
#> Arctocephalus_galapagoensis 0.47871747
#> Arctocephalus_gazella       0.56465339
#> Arctocephalus_philippii     0.44247109
#> Arctocephalus_pusillus      0.37884040
#> Arctocephalus_townsendi     0.64653619
#> Arctocephalus_tropicalis    0.32036900
#> Balaena_mysticetus          0.57306402
#> Balaenoptera_acutorostrata  0.41489069
#> Balaenoptera_bonaerensis    0.32978308
#> Balaenoptera_borealis       0.39976514
#> Balaenoptera_edeni          0.27554077
#> Balaenoptera_musculus       0.70171703
#> Balaenoptera_omurai         0.41958574
#> Balaenoptera_physalus       0.52385972
#> Berardius_arnuxii           0.24980784
#> Berardius_bairdii           0.23813008
#> Callorhinus_ursinus         0.60635396
#> Caperea_marginata           0.40111351
#> Cephalorhynchus_commersonii 0.17480251
#> Cephalorhynchus_eutropia    0.06626082
#> Cephalorhynchus_heavisidii  0.06542231
#> Cephalorhynchus_hectori     0.01634259
#> Cystophora_cristata         0.43824460
#> Delphinapterus_leucas       0.32847428
#> Delphinus_capensis          0.16472623
#> Delphinus_delphis           0.16127757
#> Dugong_dugon                0.30543867
#> Enhydra_lutris              0.45056295
#> Erignathus_barbatus         0.35107768
#> Eschrichtius_robustus       0.61300225
#> Eubalaena_australis         0.78268639
#> Eubalaena_glacialis         0.54593451
#> Eubalaena_japonica          0.54157645
#> Eumetopias_jubatus          0.56433150
#> Feresa_attenuata            0.27261560
#> Globicephala_macrorhynchus  0.35104810
#> Globicephala_melas          0.35763988
#> Grampus_griseus             0.29251186
#> Halichoerus_grypus          0.35586820
#> Histriophoca_fasciata       0.28824480
#> Hydrurga_leptonyx           0.44212015
#> Hyperoodon_ampullatus       0.12270144
#> Hyperoodon_planifrons       0.12249240
#> Indopacetus_pacificus       0.28554866
#> Kogia_breviceps             0.25751272
#> Kogia_sima                  0.25492545
#> Lagenodelphis_hosei         0.51940242
#> Lagenorhynchus_acutus       0.19736356
#> Lagenorhynchus_albirostris  0.14280471
#> Lagenorhynchus_australis    0.21058380
#> Lagenorhynchus_cruciger     0.15621825
#> Lagenorhynchus_obliquidens  0.20430708
#> Lagenorhynchus_obscurus     0.17545748
#> Leptonychotes_weddellii     0.55524903
#> Lissodelphis_borealis       0.50809556
#> Lissodelphis_peronii        0.43157141
#> Lobodon_carcinophaga        0.41597696
#> Lontra_felina               0.35413972
#> Megaptera_novaeangliae      0.45460052
#> Mesoplodon_bidens           0.23803164
#> Mesoplodon_bowdoini         0.20799830
#> Mesoplodon_carlhubbsi       0.07159363
#> Mesoplodon_densirostris     0.19967333
#> Mesoplodon_europaeus        0.09000617
#> Mesoplodon_ginkgodens       0.07202895
#> Mesoplodon_grayi            0.20125592
#> Mesoplodon_hectori          0.20213100
#> Mesoplodon_layardii         0.09936896
#> Mesoplodon_mirus            0.02711371
#> Mesoplodon_perrini          0.07105382
#> Mesoplodon_peruvianus       0.07148019
#> Mesoplodon_stejnegeri       0.07240253
#> Mesoplodon_traversii        0.09237153
#> Mirounga_angustirostris     1.00000000
#> Mirounga_leonina            0.62780483
#> Monachus_monachus           0.16152114
#> Monachus_schauinslandi      0.52972758
#> Monodon_monoceros           0.45028422
#> Neophoca_cinerea            0.39804367
#> Neophocaena_phocaenoides    0.17076625
#> Odobenus_rosmarus           0.45366482
#> Ommatophoca_rossii          0.34130521
#> Orcaella_brevirostris       0.04578858
#> Orcaella_heinsohni          0.08519702
#> Orcinus_orca                0.39332028
#> Otaria_flavescens           0.34394151
#> Pagophilus_groenlandicus    0.46023398
#> Peponocephala_electra       0.48063075
#> Phoca_largha                0.33477382
#> Phoca_vitulina              0.37496303
#> Phocarctos_hookeri          0.35171865
#> Phocoena_dioptrica          0.41532085
#> Phocoena_phocoena           0.00000000
#> Phocoena_sinus              0.14901742
#> Phocoena_spinipinnis        0.18774230
#> Phocoenoides_dalli          0.15983045
#> Physeter_macrocephalus      0.51249076
#> Pontoporia_blainvillei      0.22018302
#> Pseudorca_crassidens        0.24712980
#> Pusa_caspica                0.34349375
#> Pusa_hispida                0.15240645
#> Sotalia_guianensis          0.02364730
#> Sousa_chinensis             0.31006940
#> Sousa_teuszii               0.15850980
#> Stenella_attenuata          0.27718723
#> Stenella_clymene            0.40334210
#> Stenella_coeruleoalba       0.18363765
#> Stenella_frontalis          0.23845778
#> Stenella_longirostris       0.18285982
#> Steno_bredanensis           0.21805988
#> Tasmacetus_shepherdi        0.17676343
#> Trichechus_manatus          0.24035976
#> Trichechus_senegalensis     0.28350122
#> Tursiops_aduncus            0.13834135
#> Tursiops_truncatus          0.19082721
#> Ursus_maritimus             0.70012282
#> Zalophus_californianus      0.62677747
#> Zalophus_wollebaeki         0.41799919
#> Ziphius_cavirostris         0.09353647
    
 FUSE_res2 <- mFD::fuse(
    sp_dist        = sp_dist_tr, 
    sp_faxes_coord = as.matrix(sp_faxes_coord), 
    nb_NN          = 5,
    GE             = sp_tr$IUCN_50,
    standGE        = TRUE)
 FUSE_res2
#>                                     FUSE         FUGE         FSGE      FUn_std
#> Arctocephalus_australis     0.0000000000 0.000000e+00 0.000000e+00 0.4066626897
#> Arctocephalus_forsteri      0.0000000000 0.000000e+00 0.000000e+00 0.5170579368
#> Arctocephalus_galapagoensis 0.4051935747 2.168355e-01 1.883581e-01 0.5592652556
#> Arctocephalus_gazella       0.0000000000 0.000000e+00 0.000000e+00 0.6208798411
#> Arctocephalus_philippii     0.0000000000 0.000000e+00 0.000000e+00 0.2779525182
#> Arctocephalus_pusillus      0.0000000000 0.000000e+00 0.000000e+00 0.2442786896
#> Arctocephalus_townsendi     0.0000000000 0.000000e+00 0.000000e+00 0.5852680899
#> Arctocephalus_tropicalis    0.0000000000 0.000000e+00 0.000000e+00 0.2297455947
#> Balaena_mysticetus          0.0000000000 0.000000e+00 0.000000e+00 0.7452467855
#> Balaenoptera_acutorostrata  0.0000000000 0.000000e+00 0.000000e+00 0.4460157412
#> Balaenoptera_bonaerensis    0.0022311864 8.890871e-04 1.342099e-03 0.2184186166
#> Balaenoptera_borealis       0.2482137302 8.857883e-02 1.596349e-01 0.2139234148
#> Balaenoptera_edeni          0.0000000000 0.000000e+00 0.000000e+00 0.4661991166
#> Balaenoptera_musculus       0.5062324442 2.409373e-01 2.652951e-01 0.6292520413
#> Balaenoptera_omurai                   NA           NA           NA 0.2386976599
#> Balaenoptera_physalus       0.3468953473 1.424776e-01 2.044177e-01 0.3536749038
#> Berardius_arnuxii                     NA           NA           NA 0.3691785951
#> Berardius_bairdii                     NA           NA           NA 0.3306752012
#> Callorhinus_ursinus         0.0562359308 2.548782e-02 3.074811e-02 0.5012945401
#> Caperea_marginata           0.0000000000 0.000000e+00 0.000000e+00 0.1832726206
#> Cephalorhynchus_commersonii 0.0000000000 0.000000e+00 0.000000e+00 0.1835307609
#> Cephalorhynchus_eutropia    0.0002983037 2.850115e-05 2.698025e-04 0.0069987566
#> Cephalorhynchus_heavisidii  0.0003002586 3.386988e-05 2.663887e-04 0.0083171263
#> Cephalorhynchus_hectori     0.0101183562 3.067578e-03 7.050779e-03 0.0070959999
#> Cystophora_cristata         0.0520745374 2.975694e-02 2.231760e-02 0.5865158316
#> Delphinapterus_leucas       0.0000000000 0.000000e+00 0.000000e+00 0.6739629047
#> Delphinus_capensis                    NA           NA           NA 0.0264311644
#> Delphinus_delphis           0.0000000000 0.000000e+00 0.000000e+00 0.3984163792
#> Dugong_dugon                0.0429309730 2.732407e-02 1.560690e-02 0.5379056252
#> Enhydra_lutris              0.4434975661 2.652878e-01 1.782097e-01 0.7016950229
#> Erignathus_barbatus         0.0000000000 0.000000e+00 0.000000e+00 0.5531416419
#> Eschrichtius_robustus       0.0000000000 0.000000e+00 0.000000e+00 0.8271900919
#> Eubalaena_australis         0.0000000000 0.000000e+00 0.000000e+00 0.3144018773
#> Eubalaena_glacialis         0.3270161133 1.148380e-01 2.121781e-01 0.2810689133
#> Eubalaena_japonica          0.3945009065 1.838501e-01 2.106508e-01 0.4661757614
#> Eumetopias_jubatus          0.0041331511 1.837618e-03 2.295533e-03 0.4516548620
#> Feresa_attenuata            0.0000000000 0.000000e+00 0.000000e+00 0.2955099743
#> Globicephala_macrorhynchus  0.0000000000 0.000000e+00 0.000000e+00 0.5255589486
#> Globicephala_melas          0.0000000000 0.000000e+00 0.000000e+00 0.4990595861
#> Grampus_griseus             0.0000000000 0.000000e+00 0.000000e+00 0.5233269607
#> Halichoerus_grypus          0.0000000000 0.000000e+00 0.000000e+00 0.1032026841
#> Histriophoca_fasciata       0.0000000000 0.000000e+00 0.000000e+00 0.6521823871
#> Hydrurga_leptonyx           0.0000000000 0.000000e+00 0.000000e+00 0.6622353793
#> Hyperoodon_ampullatus                 NA           NA           NA 0.2547878402
#> Hyperoodon_planifrons       0.0000000000 0.000000e+00 0.000000e+00 0.2557952131
#> Indopacetus_pacificus                 NA           NA           NA 0.2850555343
#> Kogia_breviceps                       NA           NA           NA 0.2839209537
#> Kogia_sima                            NA           NA           NA 0.1456685909
#> Lagenodelphis_hosei         0.0000000000 0.000000e+00 0.000000e+00 0.7241772659
#> Lagenorhynchus_acutus       0.0000000000 0.000000e+00 0.000000e+00 0.4544319444
#> Lagenorhynchus_albirostris  0.0000000000 0.000000e+00 0.000000e+00 0.0907709847
#> Lagenorhynchus_australis    0.0000000000 0.000000e+00 0.000000e+00 0.0010393047
#> Lagenorhynchus_cruciger     0.0000000000 0.000000e+00 0.000000e+00 0.0020371106
#> Lagenorhynchus_obliquidens  0.0000000000 0.000000e+00 0.000000e+00 0.3327438642
#> Lagenorhynchus_obscurus     0.0000000000 0.000000e+00 0.000000e+00 0.4240208007
#> Leptonychotes_weddellii     0.0000000000 0.000000e+00 0.000000e+00 0.5404210014
#> Lissodelphis_borealis       0.0000000000 0.000000e+00 0.000000e+00 0.5812804661
#> Lissodelphis_peronii        0.0000000000 0.000000e+00 0.000000e+00 0.2947422923
#> Lobodon_carcinophaga        0.0000000000 0.000000e+00 0.000000e+00 0.5833807958
#> Lontra_felina               0.3649494018 2.222973e-01 1.426521e-01 0.5749777422
#> Megaptera_novaeangliae      0.0000000000 0.000000e+00 0.000000e+00 0.3564239828
#> Mesoplodon_bidens                     NA           NA           NA 0.2401439269
#> Mesoplodon_bowdoini                   NA           NA           NA 0.0644576583
#> Mesoplodon_carlhubbsi                 NA           NA           NA 0.0008237359
#> Mesoplodon_densirostris               NA           NA           NA 0.2058523276
#> Mesoplodon_europaeus                  NA           NA           NA 0.1358313979
#> Mesoplodon_ginkgodens                 NA           NA           NA 0.0013549553
#> Mesoplodon_grayi                      NA           NA           NA 0.0544755693
#> Mesoplodon_hectori                    NA           NA           NA 0.0533639848
#> Mesoplodon_layardii                   NA           NA           NA 0.0156247025
#> Mesoplodon_mirus                      NA           NA           NA 0.2615344649
#> Mesoplodon_perrini                    NA           NA           NA 0.0007873510
#> Mesoplodon_peruvianus                 NA           NA           NA 0.0022081471
#> Mesoplodon_stejnegeri                 NA           NA           NA 0.0019316220
#> Mesoplodon_traversii                  NA           NA           NA 0.0000000000
#> Mirounga_angustirostris     0.0000000000 0.000000e+00 0.000000e+00 1.0000000000
#> Mirounga_leonina            0.0000000000 0.000000e+00 0.000000e+00 0.6948598959
#> Monachus_monachus           0.1497520468 8.215670e-02 6.759534e-02 0.1977684465
#> Monachus_schauinslandi      0.4302127336 2.237263e-01 2.064865e-01 0.5791027790
#> Monodon_monoceros           0.0000000000 0.000000e+00 0.000000e+00 0.7032597866
#> Neophoca_cinerea            0.2192750379 6.027570e-02 1.589993e-01 0.1434988631
#> Neophocaena_phocaenoides    0.0362881475 2.753255e-02 8.755593e-03 0.5420666711
#> Odobenus_rosmarus                     NA           NA           NA 0.5803602059
#> Ommatophoca_rossii          0.0000000000 0.000000e+00 0.000000e+00 0.6659562055
#> Orcaella_brevirostris       0.0240062354 4.375540e-03 1.963070e-02 0.0101282373
#> Orcaella_heinsohni          0.0058839286 1.506092e-03 4.377837e-03 0.0292679517
#> Orcinus_orca                          NA           NA           NA 0.7536273607
#> Otaria_flavescens           0.0000000000 0.000000e+00 0.000000e+00 0.0808815948
#> Pagophilus_groenlandicus    0.0000000000 0.000000e+00 0.000000e+00 0.7570041744
#> Peponocephala_electra       0.0000000000 0.000000e+00 0.000000e+00 0.5314927079
#> Phoca_largha                0.0000000000 0.000000e+00 0.000000e+00 0.5982829480
#> Phoca_vitulina              0.0000000000 0.000000e+00 0.000000e+00 0.4543455128
#> Phocarctos_hookeri          0.1750436555 3.330083e-02 1.417428e-01 0.0782092491
#> Phocoena_dioptrica          0.0000000000 0.000000e+00 0.000000e+00 0.2513166389
#> Phocoena_phocoena           0.0000000000 0.000000e+00 0.000000e+00 0.0736635094
#> Phocoena_sinus              0.3098959997 1.709888e-01 1.389072e-01 0.1864775088
#> Phocoena_spinipinnis        0.0018432814 1.079017e-03 7.642649e-04 0.2651030237
#> Phocoenoides_dalli          0.0000000000 0.000000e+00 0.000000e+00 0.2916871646
#> Physeter_macrocephalus      0.0642642678 3.821453e-02 2.604973e-02 0.7564270292
#> Pontoporia_blainvillei      0.0302045496 1.892948e-02 1.127507e-02 0.3710816347
#> Pseudorca_crassidens        0.0029361868 1.930288e-03 1.005899e-03 0.4744533574
#> Pusa_caspica                0.2193438717 8.069628e-02 1.386476e-01 0.1941091818
#> Pusa_hispida                0.0000000000 0.000000e+00 0.000000e+00 0.1823213934
#> Sotalia_guianensis          0.0003188826 2.225865e-04 9.629603e-05 0.0546637568
#> Sousa_chinensis             0.0035493076 2.287385e-03 1.261922e-03 0.5623262797
#> Sousa_teuszii               0.3882639834 2.411295e-01 1.471345e-01 0.2726857890
#> Stenella_attenuata          0.0000000000 0.000000e+00 0.000000e+00 0.4500796407
#> Stenella_clymene            0.0000000000 0.000000e+00 0.000000e+00 0.3747462296
#> Stenella_coeruleoalba       0.0000000000 0.000000e+00 0.000000e+00 0.2311412497
#> Stenella_frontalis          0.0000000000 0.000000e+00 0.000000e+00 0.3490478318
#> Stenella_longirostris       0.0000000000 0.000000e+00 0.000000e+00 0.5133002403
#> Steno_bredanensis           0.0000000000 0.000000e+00 0.000000e+00 0.1334810508
#> Tasmacetus_shepherdi                  NA           NA           NA 0.3474343349
#> Trichechus_manatus          0.0379151532 2.561321e-02 1.230195e-02 0.5037922863
#> Trichechus_senegalensis     0.0414988099 2.700475e-02 1.449406e-02 0.5315342856
#> Tursiops_aduncus            0.0007741038 2.108846e-04 5.632192e-04 0.0517896426
#> Tursiops_truncatus          0.0000000000 0.000000e+00 0.000000e+00 0.3052280508
#> Ursus_maritimus             0.0811703393 4.575052e-02 3.541982e-02 0.9090386502
#> Zalophus_californianus      0.0000000000 0.000000e+00 0.000000e+00 0.1204667495
#> Zalophus_wollebaeki         0.1982794545 3.193729e-02 1.663422e-01 0.0749554707
#> Ziphius_cavirostris         0.0000000000 0.000000e+00 0.000000e+00 0.1383689569
#>                                FSp_std
#> Arctocephalus_australis     0.44574291
#> Arctocephalus_forsteri      0.51871477
#> Arctocephalus_galapagoensis 0.47871747
#> Arctocephalus_gazella       0.56465339
#> Arctocephalus_philippii     0.44247109
#> Arctocephalus_pusillus      0.37884040
#> Arctocephalus_townsendi     0.64653619
#> Arctocephalus_tropicalis    0.32036900
#> Balaena_mysticetus          0.57306402
#> Balaenoptera_acutorostrata  0.41489069
#> Balaenoptera_bonaerensis    0.32978308
#> Balaenoptera_borealis       0.39976514
#> Balaenoptera_edeni          0.27554077
#> Balaenoptera_musculus       0.70171703
#> Balaenoptera_omurai         0.41958574
#> Balaenoptera_physalus       0.52385972
#> Berardius_arnuxii           0.24980784
#> Berardius_bairdii           0.23813008
#> Callorhinus_ursinus         0.60635396
#> Caperea_marginata           0.40111351
#> Cephalorhynchus_commersonii 0.17480251
#> Cephalorhynchus_eutropia    0.06626082
#> Cephalorhynchus_heavisidii  0.06542231
#> Cephalorhynchus_hectori     0.01634259
#> Cystophora_cristata         0.43824460
#> Delphinapterus_leucas       0.32847428
#> Delphinus_capensis          0.16472623
#> Delphinus_delphis           0.16127757
#> Dugong_dugon                0.30543867
#> Enhydra_lutris              0.45056295
#> Erignathus_barbatus         0.35107768
#> Eschrichtius_robustus       0.61300225
#> Eubalaena_australis         0.78268639
#> Eubalaena_glacialis         0.54593451
#> Eubalaena_japonica          0.54157645
#> Eumetopias_jubatus          0.56433150
#> Feresa_attenuata            0.27261560
#> Globicephala_macrorhynchus  0.35104810
#> Globicephala_melas          0.35763988
#> Grampus_griseus             0.29251186
#> Halichoerus_grypus          0.35586820
#> Histriophoca_fasciata       0.28824480
#> Hydrurga_leptonyx           0.44212015
#> Hyperoodon_ampullatus       0.12270144
#> Hyperoodon_planifrons       0.12249240
#> Indopacetus_pacificus       0.28554866
#> Kogia_breviceps             0.25751272
#> Kogia_sima                  0.25492545
#> Lagenodelphis_hosei         0.51940242
#> Lagenorhynchus_acutus       0.19736356
#> Lagenorhynchus_albirostris  0.14280471
#> Lagenorhynchus_australis    0.21058380
#> Lagenorhynchus_cruciger     0.15621825
#> Lagenorhynchus_obliquidens  0.20430708
#> Lagenorhynchus_obscurus     0.17545748
#> Leptonychotes_weddellii     0.55524903
#> Lissodelphis_borealis       0.50809556
#> Lissodelphis_peronii        0.43157141
#> Lobodon_carcinophaga        0.41597696
#> Lontra_felina               0.35413972
#> Megaptera_novaeangliae      0.45460052
#> Mesoplodon_bidens           0.23803164
#> Mesoplodon_bowdoini         0.20799830
#> Mesoplodon_carlhubbsi       0.07159363
#> Mesoplodon_densirostris     0.19967333
#> Mesoplodon_europaeus        0.09000617
#> Mesoplodon_ginkgodens       0.07202895
#> Mesoplodon_grayi            0.20125592
#> Mesoplodon_hectori          0.20213100
#> Mesoplodon_layardii         0.09936896
#> Mesoplodon_mirus            0.02711371
#> Mesoplodon_perrini          0.07105382
#> Mesoplodon_peruvianus       0.07148019
#> Mesoplodon_stejnegeri       0.07240253
#> Mesoplodon_traversii        0.09237153
#> Mirounga_angustirostris     1.00000000
#> Mirounga_leonina            0.62780483
#> Monachus_monachus           0.16152114
#> Monachus_schauinslandi      0.52972758
#> Monodon_monoceros           0.45028422
#> Neophoca_cinerea            0.39804367
#> Neophocaena_phocaenoides    0.17076625
#> Odobenus_rosmarus           0.45366482
#> Ommatophoca_rossii          0.34130521
#> Orcaella_brevirostris       0.04578858
#> Orcaella_heinsohni          0.08519702
#> Orcinus_orca                0.39332028
#> Otaria_flavescens           0.34394151
#> Pagophilus_groenlandicus    0.46023398
#> Peponocephala_electra       0.48063075
#> Phoca_largha                0.33477382
#> Phoca_vitulina              0.37496303
#> Phocarctos_hookeri          0.35171865
#> Phocoena_dioptrica          0.41532085
#> Phocoena_phocoena           0.00000000
#> Phocoena_sinus              0.14901742
#> Phocoena_spinipinnis        0.18774230
#> Phocoenoides_dalli          0.15983045
#> Physeter_macrocephalus      0.51249076
#> Pontoporia_blainvillei      0.22018302
#> Pseudorca_crassidens        0.24712980
#> Pusa_caspica                0.34349375
#> Pusa_hispida                0.15240645
#> Sotalia_guianensis          0.02364730
#> Sousa_chinensis             0.31006940
#> Sousa_teuszii               0.15850980
#> Stenella_attenuata          0.27718723
#> Stenella_clymene            0.40334210
#> Stenella_coeruleoalba       0.18363765
#> Stenella_frontalis          0.23845778
#> Stenella_longirostris       0.18285982
#> Steno_bredanensis           0.21805988
#> Tasmacetus_shepherdi        0.17676343
#> Trichechus_manatus          0.24035976
#> Trichechus_senegalensis     0.28350122
#> Tursiops_aduncus            0.13834135
#> Tursiops_truncatus          0.19082721
#> Ursus_maritimus             0.70012282
#> Zalophus_californianus      0.62677747
#> Zalophus_wollebaeki         0.41799919
#> Ziphius_cavirostris         0.09353647
    
 FUSE_res3 <- mFD::fuse(
    sp_dist        = sp_dist_tr, 
    sp_faxes_coord = as.matrix(sp_faxes_coord), 
    nb_NN          = 5, 
    GE             = sp_tr$IUCN_100,
    standGE        = TRUE)
 FUSE_res3
#>                                     FUSE         FUGE         FSGE      FUn_std
#> Arctocephalus_australis     0.0000000000 0.000000e+00 0.0000000000 0.4066626897
#> Arctocephalus_forsteri      0.0000000000 0.000000e+00 0.0000000000 0.5170579368
#> Arctocephalus_galapagoensis 0.5992815838 3.197471e-01 0.2795345282 0.5592652556
#> Arctocephalus_gazella       0.0000000000 0.000000e+00 0.0000000000 0.6208798411
#> Arctocephalus_philippii     0.0000000000 0.000000e+00 0.0000000000 0.2779525182
#> Arctocephalus_pusillus      0.0000000000 0.000000e+00 0.0000000000 0.2442786896
#> Arctocephalus_townsendi     0.0000000000 0.000000e+00 0.0000000000 0.5852680899
#> Arctocephalus_tropicalis    0.0000000000 0.000000e+00 0.0000000000 0.2297455947
#> Balaena_mysticetus          0.0000000000 0.000000e+00 0.0000000000 0.7452467855
#> Balaenoptera_acutorostrata  0.0000000000 0.000000e+00 0.0000000000 0.4460157412
#> Balaenoptera_bonaerensis    0.0054747614 2.182024e-03 0.0032927369 0.2184186166
#> Balaenoptera_borealis       0.3731209005 1.346368e-01 0.2384841137 0.2139234148
#> Balaenoptera_edeni          0.0000000000 0.000000e+00 0.0000000000 0.4661991166
#> Balaenoptera_musculus       0.7405518753 3.534206e-01 0.3871313064 0.6292520413
#> Balaenoptera_omurai                   NA           NA           NA 0.2386976599
#> Balaenoptera_physalus       0.5159872370 2.137171e-01 0.3022700999 0.3536749038
#> Berardius_arnuxii                     NA           NA           NA 0.3691785951
#> Berardius_bairdii                     NA           NA           NA 0.3306752012
#> Callorhinus_ursinus         0.1087457494 4.935218e-02 0.0593935656 0.5012945401
#> Caperea_marginata           0.0000000000 0.000000e+00 0.0000000000 0.1832726206
#> Cephalorhynchus_commersonii 0.0000000000 0.000000e+00 0.0000000000 0.1835307609
#> Cephalorhynchus_eutropia    0.0007324479 6.999219e-05 0.0006624557 0.0069987566
#> Cephalorhynchus_heavisidii  0.0007372515 8.317621e-05 0.0006540753 0.0083171263
#> Cephalorhynchus_hectori     0.0157191188 4.769216e-03 0.0109499032 0.0070959999
#> Cystophora_cristata         0.1007823971 5.750518e-02 0.0432772161 0.5865158316
#> Delphinapterus_leucas       0.0000000000 0.000000e+00 0.0000000000 0.6739629047
#> Delphinus_capensis                    NA           NA           NA 0.0264311644
#> Delphinus_delphis           0.0000000000 0.000000e+00 0.0000000000 0.3984163792
#> Dugong_dugon                0.0832219750 5.286286e-02 0.0303591146 0.5379056252
#> Enhydra_lutris              0.6522096715 3.871212e-01 0.2650884317 0.7016950229
#> Erignathus_barbatus         0.0000000000 0.000000e+00 0.0000000000 0.5531416419
#> Eschrichtius_robustus       0.0000000000 0.000000e+00 0.0000000000 0.8271900919
#> Eubalaena_australis         0.0000000000 0.000000e+00 0.0000000000 0.3144018773
#> Eubalaena_glacialis         0.4866156416 1.734131e-01 0.3132025075 0.2810689133
#> Eubalaena_japonica          0.5841787823 2.731251e-01 0.3110536485 0.4661757614
#> Eumetopias_jubatus          0.0101348519 4.506834e-03 0.0056280180 0.4516548620
#> Feresa_attenuata            0.0000000000 0.000000e+00 0.0000000000 0.2955099743
#> Globicephala_macrorhynchus  0.0000000000 0.000000e+00 0.0000000000 0.5255589486
#> Globicephala_melas          0.0000000000 0.000000e+00 0.0000000000 0.4990595861
#> Grampus_griseus             0.0000000000 0.000000e+00 0.0000000000 0.5233269607
#> Halichoerus_grypus          0.0000000000 0.000000e+00 0.0000000000 0.1032026841
#> Histriophoca_fasciata       0.0000000000 0.000000e+00 0.0000000000 0.6521823871
#> Hydrurga_leptonyx           0.0000000000 0.000000e+00 0.0000000000 0.6622353793
#> Hyperoodon_ampullatus                 NA           NA           NA 0.2547878402
#> Hyperoodon_planifrons       0.0000000000 0.000000e+00 0.0000000000 0.2557952131
#> Indopacetus_pacificus                 NA           NA           NA 0.2850555343
#> Kogia_breviceps                       NA           NA           NA 0.2839209537
#> Kogia_sima                            NA           NA           NA 0.1456685909
#> Lagenodelphis_hosei         0.0000000000 0.000000e+00 0.0000000000 0.7241772659
#> Lagenorhynchus_acutus       0.0000000000 0.000000e+00 0.0000000000 0.4544319444
#> Lagenorhynchus_albirostris  0.0000000000 0.000000e+00 0.0000000000 0.0907709847
#> Lagenorhynchus_australis    0.0000000000 0.000000e+00 0.0000000000 0.0010393047
#> Lagenorhynchus_cruciger     0.0000000000 0.000000e+00 0.0000000000 0.0020371106
#> Lagenorhynchus_obliquidens  0.0000000000 0.000000e+00 0.0000000000 0.3327438642
#> Lagenorhynchus_obscurus     0.0000000000 0.000000e+00 0.0000000000 0.4240208007
#> Leptonychotes_weddellii     0.0000000000 0.000000e+00 0.0000000000 0.5404210014
#> Lissodelphis_borealis       0.0000000000 0.000000e+00 0.0000000000 0.5812804661
#> Lissodelphis_peronii        0.0000000000 0.000000e+00 0.0000000000 0.2947422923
#> Lobodon_carcinophaga        0.0000000000 0.000000e+00 0.0000000000 0.5833807958
#> Lontra_felina               0.5413762905 3.274063e-01 0.2139699947 0.5749777422
#> Megaptera_novaeangliae      0.0000000000 0.000000e+00 0.0000000000 0.3564239828
#> Mesoplodon_bidens                     NA           NA           NA 0.2401439269
#> Mesoplodon_bowdoini                   NA           NA           NA 0.0644576583
#> Mesoplodon_carlhubbsi                 NA           NA           NA 0.0008237359
#> Mesoplodon_densirostris               NA           NA           NA 0.2058523276
#> Mesoplodon_europaeus                  NA           NA           NA 0.1358313979
#> Mesoplodon_ginkgodens                 NA           NA           NA 0.0013549553
#> Mesoplodon_grayi                      NA           NA           NA 0.0544755693
#> Mesoplodon_hectori                    NA           NA           NA 0.0533639848
#> Mesoplodon_layardii                   NA           NA           NA 0.0156247025
#> Mesoplodon_mirus                      NA           NA           NA 0.2615344649
#> Mesoplodon_perrini                    NA           NA           NA 0.0007873510
#> Mesoplodon_peruvianus                 NA           NA           NA 0.0022081471
#> Mesoplodon_stejnegeri                 NA           NA           NA 0.0019316220
#> Mesoplodon_traversii                  NA           NA           NA 0.0000000000
#> Mirounga_angustirostris     0.0000000000 0.000000e+00 0.0000000000 1.0000000000
#> Mirounga_leonina            0.0000000000 0.000000e+00 0.0000000000 0.6948598959
#> Monachus_monachus           0.2283726931 1.250786e-01 0.1032941342 0.1977684465
#> Monachus_schauinslandi      0.6345952150 3.294074e-01 0.3051878051 0.5791027790
#> Monodon_monoceros           0.0000000000 0.000000e+00 0.0000000000 0.7032597866
#> Neophoca_cinerea            0.3298536309 9.228362e-02 0.2375700105 0.1434988631
#> Neophocaena_phocaenoides    0.0703478821 5.326109e-02 0.0170867934 0.5420666711
#> Odobenus_rosmarus                     NA           NA           NA 0.5803602059
#> Ommatophoca_rossii          0.0000000000 0.000000e+00 0.0000000000 0.6659562055
#> Orcaella_brevirostris       0.0371819980 6.800264e-03 0.0303817341 0.0101282373
#> Orcaella_heinsohni          0.0115106174 2.949347e-03 0.0085612703 0.0292679517
#> Orcinus_orca                          NA           NA           NA 0.7536273607
#> Otaria_flavescens           0.0000000000 0.000000e+00 0.0000000000 0.0808815948
#> Pagophilus_groenlandicus    0.0000000000 0.000000e+00 0.0000000000 0.7570041744
#> Peponocephala_electra       0.0000000000 0.000000e+00 0.0000000000 0.5314927079
#> Phoca_largha                0.0000000000 0.000000e+00 0.0000000000 0.5982829480
#> Phoca_vitulina              0.0000000000 0.000000e+00 0.0000000000 0.4543455128
#> Phocarctos_hookeri          0.2640009489 5.134871e-02 0.2126522354 0.0782092491
#> Phocoena_dioptrica          0.0000000000 0.000000e+00 0.0000000000 0.2513166389
#> Phocoena_phocoena           0.0000000000 0.000000e+00 0.0000000000 0.0736635094
#> Phocoena_sinus              0.3098959997 1.709888e-01 0.1389071589 0.1864775088
#> Phocoena_spinipinnis        0.0045236417 2.647790e-03 0.0018758522 0.2651030237
#> Phocoenoides_dalli          0.0000000000 0.000000e+00 0.0000000000 0.2916871646
#> Physeter_macrocephalus      0.1239917183 7.356461e-02 0.0504271100 0.7564270292
#> Pontoporia_blainvillei      0.0587425276 3.676510e-02 0.0219774297 0.3710816347
#> Pseudorca_crassidens        0.0072022892 4.733791e-03 0.0024684984 0.4744533574
#> Pusa_caspica                0.3310632694 1.229008e-01 0.2081624960 0.1941091818
#> Pusa_hispida                0.0000000000 0.000000e+00 0.0000000000 0.1823213934
#> Sotalia_guianensis          0.0007830123 5.465434e-04 0.0002364689 0.0546637568
#> Sousa_chinensis             0.0087042852 5.608076e-03 0.0030962090 0.5623262797
#> Sousa_teuszii               0.3882639834 2.411295e-01 0.1471345215 0.2726857890
#> Stenella_attenuata          0.0000000000 0.000000e+00 0.0000000000 0.4500796407
#> Stenella_clymene            0.0000000000 0.000000e+00 0.0000000000 0.3747462296
#> Stenella_coeruleoalba       0.0000000000 0.000000e+00 0.0000000000 0.2311412497
#> Stenella_frontalis          0.0000000000 0.000000e+00 0.0000000000 0.3490478318
#> Stenella_longirostris       0.0000000000 0.000000e+00 0.0000000000 0.5133002403
#> Steno_bredanensis           0.0000000000 0.000000e+00 0.0000000000 0.1334810508
#> Tasmacetus_shepherdi                  NA           NA           NA 0.3474343349
#> Trichechus_manatus          0.0735594963 4.959209e-02 0.0239674088 0.5037922863
#> Trichechus_senegalensis     0.0804618871 5.225279e-02 0.0282090962 0.5315342856
#> Tursiops_aduncus            0.0019004117 5.178147e-04 0.0013825971 0.0517896426
#> Tursiops_truncatus          0.0000000000 0.000000e+00 0.0000000000 0.3052280508
#> Ursus_maritimus             0.1560437697 8.777232e-02 0.0682714461 0.9090386502
#> Zalophus_californianus      0.0000000000 0.000000e+00 0.0000000000 0.1204667495
#> Zalophus_wollebaeki         0.2973796407 4.926418e-02 0.2481154637 0.0749554707
#> Ziphius_cavirostris         0.0000000000 0.000000e+00 0.0000000000 0.1383689569
#>                                FSp_std
#> Arctocephalus_australis     0.44574291
#> Arctocephalus_forsteri      0.51871477
#> Arctocephalus_galapagoensis 0.47871747
#> Arctocephalus_gazella       0.56465339
#> Arctocephalus_philippii     0.44247109
#> Arctocephalus_pusillus      0.37884040
#> Arctocephalus_townsendi     0.64653619
#> Arctocephalus_tropicalis    0.32036900
#> Balaena_mysticetus          0.57306402
#> Balaenoptera_acutorostrata  0.41489069
#> Balaenoptera_bonaerensis    0.32978308
#> Balaenoptera_borealis       0.39976514
#> Balaenoptera_edeni          0.27554077
#> Balaenoptera_musculus       0.70171703
#> Balaenoptera_omurai         0.41958574
#> Balaenoptera_physalus       0.52385972
#> Berardius_arnuxii           0.24980784
#> Berardius_bairdii           0.23813008
#> Callorhinus_ursinus         0.60635396
#> Caperea_marginata           0.40111351
#> Cephalorhynchus_commersonii 0.17480251
#> Cephalorhynchus_eutropia    0.06626082
#> Cephalorhynchus_heavisidii  0.06542231
#> Cephalorhynchus_hectori     0.01634259
#> Cystophora_cristata         0.43824460
#> Delphinapterus_leucas       0.32847428
#> Delphinus_capensis          0.16472623
#> Delphinus_delphis           0.16127757
#> Dugong_dugon                0.30543867
#> Enhydra_lutris              0.45056295
#> Erignathus_barbatus         0.35107768
#> Eschrichtius_robustus       0.61300225
#> Eubalaena_australis         0.78268639
#> Eubalaena_glacialis         0.54593451
#> Eubalaena_japonica          0.54157645
#> Eumetopias_jubatus          0.56433150
#> Feresa_attenuata            0.27261560
#> Globicephala_macrorhynchus  0.35104810
#> Globicephala_melas          0.35763988
#> Grampus_griseus             0.29251186
#> Halichoerus_grypus          0.35586820
#> Histriophoca_fasciata       0.28824480
#> Hydrurga_leptonyx           0.44212015
#> Hyperoodon_ampullatus       0.12270144
#> Hyperoodon_planifrons       0.12249240
#> Indopacetus_pacificus       0.28554866
#> Kogia_breviceps             0.25751272
#> Kogia_sima                  0.25492545
#> Lagenodelphis_hosei         0.51940242
#> Lagenorhynchus_acutus       0.19736356
#> Lagenorhynchus_albirostris  0.14280471
#> Lagenorhynchus_australis    0.21058380
#> Lagenorhynchus_cruciger     0.15621825
#> Lagenorhynchus_obliquidens  0.20430708
#> Lagenorhynchus_obscurus     0.17545748
#> Leptonychotes_weddellii     0.55524903
#> Lissodelphis_borealis       0.50809556
#> Lissodelphis_peronii        0.43157141
#> Lobodon_carcinophaga        0.41597696
#> Lontra_felina               0.35413972
#> Megaptera_novaeangliae      0.45460052
#> Mesoplodon_bidens           0.23803164
#> Mesoplodon_bowdoini         0.20799830
#> Mesoplodon_carlhubbsi       0.07159363
#> Mesoplodon_densirostris     0.19967333
#> Mesoplodon_europaeus        0.09000617
#> Mesoplodon_ginkgodens       0.07202895
#> Mesoplodon_grayi            0.20125592
#> Mesoplodon_hectori          0.20213100
#> Mesoplodon_layardii         0.09936896
#> Mesoplodon_mirus            0.02711371
#> Mesoplodon_perrini          0.07105382
#> Mesoplodon_peruvianus       0.07148019
#> Mesoplodon_stejnegeri       0.07240253
#> Mesoplodon_traversii        0.09237153
#> Mirounga_angustirostris     1.00000000
#> Mirounga_leonina            0.62780483
#> Monachus_monachus           0.16152114
#> Monachus_schauinslandi      0.52972758
#> Monodon_monoceros           0.45028422
#> Neophoca_cinerea            0.39804367
#> Neophocaena_phocaenoides    0.17076625
#> Odobenus_rosmarus           0.45366482
#> Ommatophoca_rossii          0.34130521
#> Orcaella_brevirostris       0.04578858
#> Orcaella_heinsohni          0.08519702
#> Orcinus_orca                0.39332028
#> Otaria_flavescens           0.34394151
#> Pagophilus_groenlandicus    0.46023398
#> Peponocephala_electra       0.48063075
#> Phoca_largha                0.33477382
#> Phoca_vitulina              0.37496303
#> Phocarctos_hookeri          0.35171865
#> Phocoena_dioptrica          0.41532085
#> Phocoena_phocoena           0.00000000
#> Phocoena_sinus              0.14901742
#> Phocoena_spinipinnis        0.18774230
#> Phocoenoides_dalli          0.15983045
#> Physeter_macrocephalus      0.51249076
#> Pontoporia_blainvillei      0.22018302
#> Pseudorca_crassidens        0.24712980
#> Pusa_caspica                0.34349375
#> Pusa_hispida                0.15240645
#> Sotalia_guianensis          0.02364730
#> Sousa_chinensis             0.31006940
#> Sousa_teuszii               0.15850980
#> Stenella_attenuata          0.27718723
#> Stenella_clymene            0.40334210
#> Stenella_coeruleoalba       0.18363765
#> Stenella_frontalis          0.23845778
#> Stenella_longirostris       0.18285982
#> Steno_bredanensis           0.21805988
#> Tasmacetus_shepherdi        0.17676343
#> Trichechus_manatus          0.24035976
#> Trichechus_senegalensis     0.28350122
#> Tursiops_aduncus            0.13834135
#> Tursiops_truncatus          0.19082721
#> Ursus_maritimus             0.70012282
#> Zalophus_californianus      0.62677747
#> Zalophus_wollebaeki         0.41799919
#> Ziphius_cavirostris         0.09353647